Change variable types to uint64_t

+++ src/structures/typedefs.h
@@ -30,12 +30,12 @@ namespace vroom {
 // To easily differentiate variable types.
 using Id = uint64_t;
 using Index = uint16_t;
-using Cost = uint32_t;
+using Cost = uint64_t;
 using Gain = int64_t;
-using Distance = uint32_t;
-using Duration = uint32_t;
+using Distance = uint64_t;
+using Duration = uint64_t;
 using Coordinate = double;
-using Capacity = int64_t;
+using Capacity = uint64_t;
 using Skill = uint32_t;
 using Priority = uint32_t;
